Day8: UserModel, handle Multiple Users

+++ database/category_models.py
@@ -1,17 +1,32 @@
 from database.database_manager import DatabaseManager
 import config
 from datetime import datetime
+from typing import Optional
+from bson.objectid import ObjectId
 
 collection_name = config.COLLECTIONS['category']
 
 class CategoryModel:
-    def __init__(self):
+    def __init__(self, user_id: Optional[str] = None):
         self.db_manager = DatabaseManager()
         self.collection = self.db_manager.get_collection(collection_name=collection_name)
-        self._initialize_default_categories()
 
-    def _initialize_default_categories(self):
-        """Initialize categories if they dont exist"""
+        # init:
+        self.user_id = user_id
+
+    def set_user_id(self, user_id: str):
+        self.user_id = ObjectId(user_id) if user_id is not None else None
+
+        # after we have user_id, initialize their default categories
+        self._initialize_user_default_categories()
+
+    def _initialize_user_default_categories(self):
+        """Initialize user categories if they dont exist"""
+
+        # Check if there is user_id, exist earlier
+        if not self.user_id:
+            return
+        
         # EXPENSE
         for cate in config.DEFAULT_CATEGORIES_EXPENSE:
             # # calling by params order
@@ -29,7 +44,8 @@ class CategoryModel:
         # define filter
         filter_ = {
             "type": category_type,
-            "name": category_name
+            "name": category_name,
+            "user_id": self.user_id
         }
 
         # define update_doc
@@ -50,17 +66,14 @@ class CategoryModel:
         return result.upserted_id
 
     def delete_category(self, category_type: str, category_name: str):
-        result = self.collection.delete_one({"type": category_type, "name": category_name})
+        result = self.collection.delete_one({"type": category_type, "name": category_name, "user_id": self.user_id}) # add user_id condition
         return result.deleted_count
 
-    def get_category_by_type(self, category_type: str):
-        # result = self.collection.find({"type": category_type})
-        # result = list(result)
-        # return result
-        return list(self.collection.find({"type": category_type}).sort("created_at", -1))
+    def get_categories_by_type(self, category_type: str):
+        return list(self.collection.find({"type": category_type, "user_id": self.user_id}).sort("created_at", -1))  # add user_id condition
     
     def get_total(self):
-        result = self.collection.find({})
+        result = self.collection.find({"user_id": self.user_id})
         result = list(result)
         return result
